fix: update linter config and resolve G115 and ST1005 linter errors
Update .golangci.yaml to match backend linter configuration to ensure
all appropriate linters are enabled.

- G115 - Integer overflow in pool limit:
  Change PoolLimit type from int to int32 to match pgxpool.Config.MaxConns,
  eliminating type conversion and preventing potential integer overflow.

- ST1005 - Error strings should not be capitalized
  Fix error message capitalization in zestwrapper/rpm.go

+++ pkg/tangy/interface.go
@@ -19,7 +19,7 @@ func New(dbConfig Database, logConfig Logger) (Tangy, error) {
 	if dbConfig.PoolLimit == 0 {
 		dbConfig.PoolLimit = DefaultMaxPoolLimit
 	}
-	pxConfig.MaxConns = int32(dbConfig.PoolLimit)
+	pxConfig.MaxConns = dbConfig.PoolLimit
 
 	if logConfig.Logger != nil && logConfig.Enabled {
 		zlog := zerologadapter.NewLogger(*logConfig.Logger)
